zkan/pyproject.toml
[tool.poetry]
name = "zkan"
version = "0.1.0"
description = "Very Lightweight Open Data Platform"
authors = ["Tom Wood <thomas.a.wood@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.61.0"}
motor = "^2.2.0"
python-dotenv = "^0.14.0"
jsonschema = "^3.2.0"
circuitbreaker = "^1.3.1"

# pin orjson; newer versions seem to need rust toolchain installed;
# not quite ready for that
orjson = "3.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = "^19.10b0"
pytest = "^6.0.1"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
